tech-studio-projects/financial-industry/high-frequency-trading-system-simulation/java/core-system/src/main/java/com/hft/common/ConfigLoader.java
package com.hft.common;

import java.util.Properties;

/**
 * Loads and manages application configuration, supporting runtime adjustments to avoid restarts.
 */
public class ConfigLoader {

    private Properties config;

    public ConfigLoader() {
        loadConfig("invalidConfig.properties");
    }

    public void loadConfig(String s) {
        // Load properties from file and handle caching for optimized access.
    }

    public String getConfigValue(String someKey) {
        return someKey;
    }

    public void loadConfigFromSource(String url) {
    }
}
